refactor: move more functions to parser module
